Debugging DroneEvents
DroneEvents enum was not made public and in a separate file from DroneState, i moved it to its own file as the compiler was not recognizing it in other files. There was also an issue with DroneEvents converting to and from string when parsing requests/responses, i added static functions inside DroneEvents to handle that.